requirements.txt
attrs>=19.3.0
domdf_python_tools>=0.8.1
typing_inspect>=0.6.0
